+++ test/unit/model/data_test.exs
@@ -16,7 +16,7 @@ defmodule RDF.DataTest do
         |> EX.p2(EX.O3, EX.O4)
       )
 
-    named_graph = %Graph{graph | name: iri(EX.NamedGraph)}
+    named_graph = %Graph{graph | name: RDF.iri(EX.NamedGraph)}
 
     dataset =
       Dataset.new()
@@ -114,7 +114,7 @@ defmodule RDF.DataTest do
 
     test "description when the requested subject does not match the Description.subject",
          %{description: description} do
-      assert RDF.Data.description(description, iri(EX.Other)) == Description.new(EX.Other)
+      assert RDF.Data.description(description, RDF.iri(EX.Other)) == Description.new(EX.Other)
     end
 
     test "descriptions", %{description: description} do
@@ -126,7 +126,7 @@ defmodule RDF.DataTest do
     end
 
     test "subjects", %{description: description} do
-      assert RDF.Data.subjects(description) == MapSet.new([iri(EX.S)])
+      assert RDF.Data.subjects(description) == MapSet.new([RDF.iri(EX.S)])
     end
 
     test "predicates", %{description: description} do
@@ -135,19 +135,19 @@ defmodule RDF.DataTest do
 
     test "objects", %{description: description} do
       assert RDF.Data.objects(description) ==
-               MapSet.new([iri(EX.O1), iri(EX.O2), iri(EX.O3), ~B<foo>])
+               MapSet.new([RDF.iri(EX.O1), RDF.iri(EX.O2), RDF.iri(EX.O3), ~B<foo>])
     end
 
     test "resources", %{description: description} do
       assert RDF.Data.resources(description) ==
                MapSet.new([
-                 iri(EX.S),
+                 RDF.iri(EX.S),
                  EX.p1(),
                  EX.p2(),
                  EX.p3(),
-                 iri(EX.O1),
-                 iri(EX.O2),
-                 iri(EX.O3),
+                 RDF.iri(EX.O1),
+                 RDF.iri(EX.O2),
+                 RDF.iri(EX.O3),
                  ~B<foo>
                ])
     end
@@ -282,12 +282,12 @@ defmodule RDF.DataTest do
 
     test "description when a description is present",
          %{graph: graph, description: description} do
-      assert RDF.Data.description(graph, iri(EX.S)) == description
+      assert RDF.Data.description(graph, RDF.iri(EX.S)) == description
       assert RDF.Data.description(graph, EX.S) == description
     end
 
     test "description when a description is not present", %{graph: graph} do
-      assert RDF.Data.description(graph, iri(EX.Other)) == Description.new(EX.Other)
+      assert RDF.Data.description(graph, RDF.iri(EX.Other)) == Description.new(EX.Other)
     end
 
     test "descriptions", %{graph: graph, description: description} do
@@ -300,7 +300,7 @@ defmodule RDF.DataTest do
     end
 
     test "subjects", %{graph: graph} do
-      assert RDF.Data.subjects(graph) == MapSet.new([iri(EX.S), iri(EX.S2)])
+      assert RDF.Data.subjects(graph) == MapSet.new([RDF.iri(EX.S), RDF.iri(EX.S2)])
     end
 
     test "predicates", %{graph: graph} do
@@ -309,21 +309,27 @@ defmodule RDF.DataTest do
 
     test "objects", %{graph: graph} do
       assert RDF.Data.objects(graph) ==
-               MapSet.new([iri(EX.O1), iri(EX.O2), iri(EX.O3), iri(EX.O4), ~B<foo>])
+               MapSet.new([
+                 RDF.iri(EX.O1),
+                 RDF.iri(EX.O2),
+                 RDF.iri(EX.O3),
+                 RDF.iri(EX.O4),
+                 ~B<foo>
+               ])
     end
 
     test "resources", %{graph: graph} do
       assert RDF.Data.resources(graph) ==
                MapSet.new([
-                 iri(EX.S),
-                 iri(EX.S2),
+                 RDF.iri(EX.S),
+                 RDF.iri(EX.S2),
                  EX.p1(),
                  EX.p2(),
                  EX.p3(),
-                 iri(EX.O1),
-                 iri(EX.O2),
-                 iri(EX.O3),
-                 iri(EX.O4),
+                 RDF.iri(EX.O1),
+                 RDF.iri(EX.O2),
+                 RDF.iri(EX.O3),
+                 RDF.iri(EX.O4),
                  ~B<foo>
                ])
     end
@@ -451,12 +457,12 @@ defmodule RDF.DataTest do
     test "description when a description is present",
          %{dataset: dataset, description: description} do
       description_aggregate = Description.add(description, {EX.S, EX.p3(), EX.O5})
-      assert RDF.Data.description(dataset, iri(EX.S)) == description_aggregate
+      assert RDF.Data.description(dataset, RDF.iri(EX.S)) == description_aggregate
       assert RDF.Data.description(dataset, EX.S) == description_aggregate
     end
 
     test "description when a description is not present", %{dataset: dataset} do
-      assert RDF.Data.description(dataset, iri(EX.Other)) == Description.new(EX.Other)
+      assert RDF.Data.description(dataset, RDF.iri(EX.Other)) == Description.new(EX.Other)
     end
 
     test "descriptions", %{dataset: dataset, description: description} do
@@ -474,7 +480,8 @@ defmodule RDF.DataTest do
     end
 
     test "subjects", %{dataset: dataset} do
-      assert RDF.Data.subjects(dataset) == MapSet.new([iri(EX.S), iri(EX.S2), iri(EX.S3)])
+      assert RDF.Data.subjects(dataset) ==
+               MapSet.new([RDF.iri(EX.S), RDF.iri(EX.S2), RDF.iri(EX.S3)])
     end
 
     test "predicates", %{dataset: dataset} do
@@ -483,23 +490,30 @@ defmodule RDF.DataTest do
 
     test "objects", %{dataset: dataset} do
       assert RDF.Data.objects(dataset) ==
-               MapSet.new([iri(EX.O1), iri(EX.O2), iri(EX.O3), iri(EX.O4), iri(EX.O5), ~B<foo>])
+               MapSet.new([
+                 RDF.iri(EX.O1),
+                 RDF.iri(EX.O2),
+                 RDF.iri(EX.O3),
+                 RDF.iri(EX.O4),
+                 RDF.iri(EX.O5),
+                 ~B<foo>
+               ])
     end
 
     test "resources", %{dataset: dataset} do
       assert RDF.Data.resources(dataset) ==
                MapSet.new([
-                 iri(EX.S),
-                 iri(EX.S2),
-                 iri(EX.S3),
+                 RDF.iri(EX.S),
+                 RDF.iri(EX.S2),
+                 RDF.iri(EX.S3),
                  EX.p1(),
                  EX.p2(),
                  EX.p3(),
-                 iri(EX.O1),
-                 iri(EX.O2),
-                 iri(EX.O3),
-                 iri(EX.O4),
-                 iri(EX.O5),
+                 RDF.iri(EX.O1),
+                 RDF.iri(EX.O2),
+                 RDF.iri(EX.O3),
+                 RDF.iri(EX.O4),
+                 RDF.iri(EX.O5),
                  ~B<foo>
                ])
     end